+++ CHANGELOG.md
@@ -16,8 +16,10 @@
 * Refactor Argument to Comment
 * Rename devel branch to development to be consistent with other branches
 * Remove express-promise-router dependency
+* Remove oauth userdata parsing: the API is now single source of truth
 
 ### auth server
+* Remove user.extradata: the API is now single source of truth
 
 ### image server
 * Refactor init database, merge seeds and fixtures